// web/src/api/generated/model/createBackendInputBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */
import type { CreateBackendInputBodyType } from "./createBackendInputBodyType.ts";
import type { LocalConfig } from "./localConfig.ts";
import type { S3Input } from "./s3Input.ts";

export interface CreateBackendInputBody {
  enabled: boolean;
  /** Required for local backends. */
  local?: LocalConfig;
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /** Required for S3 backends. */
  s3?: S3Input;
  type: CreateBackendInputBodyType;
}