// web/src/api/generated/model/updateBackendInputBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */
import type { LocalConfig } from "./localConfig.ts";
import type { S3Input } from "./s3Input.ts";

export interface UpdateBackendInputBody {
  enabled: boolean;
  /** Required for local backends. */
  local?: LocalConfig;
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /** Required for S3 backends. */
  s3?: S3Input;
}